# schemas/scanalyze-document-journey.openapi.v1.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Scanalyze Document Journey API",
    "version": "1.0.0",
    "description": "Authoritative additive contract for the idempotent, owner-bound Scanalyze document journey. The historical /api/v1 routes are not aliases for this contract and are not silently upgraded."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin API. Current clients call the explicit /api/v2 namespace; the /api facade remains a bounded historical /api/v1 adapter."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/api/v2/batches": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ContractVersionHeader"
        },
        {
          "$ref": "#/components/parameters/CorrelationRequestHeader"
        }
      ],
      "post": {
        "operationId": "createBatchV2",
        "summary": "Create one durable batch",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKeyHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The batch was created, or the same durable batch was replayed.",
            "headers": {
              "X-Correlation-ID": {
                "$ref": "#/components/headers/CorrelationResponseHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchCreateResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" },
          "502": { "$ref": "#/components/responses/UpstreamError" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        }
      }
    },
    "/api/v2/documents": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ContractVersionHeader"
        },
        {
          "$ref": "#/components/parameters/CorrelationRequestHeader"
        }
      ],
      "post": {
        "operationId": "createDocumentV2",
        "summary": "Create one durable upload intent",
        "description": "The durableResponse is replay-stable. uploadCapability is ephemeral, separately authorized, and is never part of the idempotency ledger. A first create (replayed=false) always includes it. An exact replay may omit it when the durable document is no longer uploadable; the client can use the owner-authorized refresh route only while the document remains uploadable.",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKeyHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The upload intent was created, or its durable projection was replayed with a fresh ephemeral capability.",
            "headers": {
              "X-Correlation-ID": {
                "$ref": "#/components/headers/CorrelationResponseHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentCreateResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" },
          "502": { "$ref": "#/components/responses/UpstreamError" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        }
      }
    },
    "/api/v2/documents/{documentId}/submit": {
      "parameters": [
        {
          "$ref": "#/components/parameters/DocumentIdPath"
        },
        {
          "$ref": "#/components/parameters/ContractVersionHeader"
        },
        {
          "$ref": "#/components/parameters/CorrelationRequestHeader"
        }
      ],
      "post": {
        "operationId": "submitDocumentV2",
        "summary": "Submit the durable document to the canonical ingest stage",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitDocumentRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Submission accepted or already enqueued.",
            "headers": {
              "X-Correlation-ID": { "$ref": "#/components/headers/CorrelationResponseHeader" }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SubmitDocumentResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" },
          "502": { "$ref": "#/components/responses/UpstreamError" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        }
      }
    },
    "/api/v2/documents/{documentId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/DocumentIdPath"
        },
        {
          "$ref": "#/components/parameters/ContractVersionHeader"
        },
        {
          "$ref": "#/components/parameters/CorrelationRequestHeader"
        }
      ],
      "get": {
        "operationId": "getDocumentStatusV2",
        "summary": "Read the closed public lifecycle projection",
        "responses": {
          "200": {
            "description": "Owner-bound document status.",
            "headers": {
              "X-Correlation-ID": { "$ref": "#/components/headers/CorrelationResponseHeader" }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DocumentStatusResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" },
          "502": { "$ref": "#/components/responses/UpstreamError" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        }
      }
    },
    "/api/v2/documents/{documentId}/upload-capabilities": {
      "parameters": [
        {
          "$ref": "#/components/parameters/DocumentIdPath"
        },
        {
          "$ref": "#/components/parameters/ContractVersionHeader"
        },
        {
          "$ref": "#/components/parameters/CorrelationRequestHeader"
        }
      ],
      "post": {
        "operationId": "refreshDocumentUploadCapabilityV2",
        "summary": "Mint a fresh bounded upload capability after owner and state checks",
        "responses": {
          "200": {
            "description": "Fresh ephemeral upload capability. This response is not durable ledger evidence.",
            "headers": {
              "X-Correlation-ID": { "$ref": "#/components/headers/CorrelationResponseHeader" }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UploadCapabilityResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" },
          "502": { "$ref": "#/components/responses/UpstreamError" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        }
      }
    },
    "/api/v2/operations/{operation}/reconciliation": {
      "parameters": [
        {
          "$ref": "#/components/parameters/OperationPath"
        },
        {
          "$ref": "#/components/parameters/ContractVersionHeader"
        },
        {
          "$ref": "#/components/parameters/CorrelationRequestHeader"
        }
      ],
      "post": {
        "operationId": "reconcileDocumentJourneyOperationV2",
        "summary": "Resolve a lost create response without a resource identifier",
        "description": "The original Idempotency-Key is scoped to the authenticated actor, customer, deployment, operation, and contract version. This lookup never creates a second business resource.",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKeyHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "Closed reconciliation state and, for SUCCEEDED, the exact durable response projection.",
            "headers": {
              "X-Correlation-ID": { "$ref": "#/components/headers/CorrelationResponseHeader" }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReconciliationResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" },
          "502": { "$ref": "#/components/responses/UpstreamError" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        }
      }
    },
    "/api/v2/documents/{documentId}/result": {
      "parameters": [
        {
          "$ref": "#/components/parameters/DocumentIdPath"
        },
        {
          "$ref": "#/components/parameters/ContractVersionHeader"
        },
        {
          "$ref": "#/components/parameters/CorrelationRequestHeader"
        }
      ],
      "get": {
        "operationId": "getDocumentResultV2",
        "summary": "Read a typed terminal result",
        "responses": {
          "200": {
            "description": "Typed, identity-bound bank_statement result. Other document families fail explicitly.",
            "headers": {
              "X-Correlation-ID": { "$ref": "#/components/headers/CorrelationResponseHeader" }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "./scanalyze-document-journey-result.v1.schema.json" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" },
          "502": { "$ref": "#/components/responses/UpstreamError" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "ContractVersionHeader": {
        "name": "X-Scanalyze-Contract-Version",
        "in": "header",
        "required": true,
        "description": "Exact document-journey contract. Missing or unsupported values are rejected; there is no downgrade.",
        "schema": {
          "const": "scanalyze.document-journey.v1"
        }
      },
      "IdempotencyKeyHeader": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Opaque canonical UUID. The raw value is never persisted or logged.",
        "schema": {
          "type": "string",
          "format": "uuid",
          "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
          "minLength": 36,
          "maxLength": 36
        }
      },
      "CorrelationRequestHeader": {
        "name": "X-Correlation-ID",
        "in": "header",
        "required": false,
        "description": "Optional opaque diagnostic reference. It never establishes identity or ownership.",
        "schema": {
          "$ref": "#/components/schemas/CorrelationReference"
        }
      },
      "DocumentIdPath": {
        "name": "documentId",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/DocumentId"
        }
      },
...
    },
    "headers": {
      "CorrelationResponseHeader": {
        "required": true,
        "description": "Safe opaque diagnostic reference for support correlation.",
        "schema": {
          "$ref": "#/components/schemas/CorrelationReference"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed JSON, duplicate keys, invalid headers, or unsupported contract version.",
        "headers": {
          "X-Correlation-ID": { "$ref": "#/components/headers/CorrelationResponseHeader" }
        },
        "content": { "application/json": { "schema": { "allOf": [
          { "$ref": "#/components/schemas/ErrorEnvelope" },
          { "properties": { "code": { "enum": ["MALFORMED_REQUEST", "UNSUPPORTED_CONTRACT_VERSION"] } }, "required": ["code"] }
        ] } } }
      },
      "Unauthorized": {
        "description": "Authentication is missing or invalid.",
        "headers": {
          "X-Correlation-ID": { "$ref": "#/components/headers/CorrelationResponseHeader" }
        },
        "content": { "application/json": { "schema": { "allOf": [
          { "$ref": "#/components/schemas/ErrorEnvelope" },
          { "properties": { "code": { "enum": ["AUTHENTICATION_REQUIRED", "AUTHENTICATION_INVALID"] } }, "required": ["code"] }
        ] } } }
      },
      "Forbidden": {
        "description": "The authenticated principal is not authorized.",
        "headers": {
          "X-Correlation-ID": { "$ref": "#/components/headers/CorrelationResponseHeader" }
        },
        "content": { "application/json": { "schema": { "allOf": [
          { "$ref": "#/components/schemas/ErrorEnvelope" },
          { "properties": { "code": { "const": "AUTHORIZATION_DENIED" } }, "required": ["code"] }
        ] } } }
      },
...
      "Conflict": {
        "description": "Idempotency conflict, result not ready, wrong document state, or expired operation key.",
        "headers": {
          "X-Correlation-ID": { "$ref": "#/components/headers/CorrelationResponseHeader" }
        },
        "content": { "application/json": { "schema": { "allOf": [
          { "$ref": "#/components/schemas/ErrorEnvelope" },
          { "properties": { "code": { "enum": ["IDEMPOTENCY_CONFLICT", "RESULT_NOT_READY", "STATE_CONFLICT", "EXPIRED_OPERATION"] } }, "required": ["code"] }
        ] } } }
      },
...
    },
    "schemas": {
...
      "BatchCreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "maxProperties": 0
      },
...
      "BatchCreateResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schemaVersion",
          "contractVersion",
          "replayed",
          "durableResponse"
        ],
        "properties": {
          "schemaVersion": { "const": "scanalyze.operation-response.v1" },
          "contractVersion": { "const": "scanalyze.document-journey.v1" },
          "replayed": { "type": "boolean" },
          "durableResponse": { "$ref": "#/components/schemas/BatchDurableResponse" }
        }
      },
...
      "SubmitDocumentRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "stage": {
            "const": "ingest",
            "default": "ingest"
          }
        }
      },
...
    }
  }
}